+++ src/main/java/com/example/tokoindah/controller/adminPage/product/EditProductController.java
@@ -51,7 +51,7 @@ public class EditProductController {
         edit_btn.setOnAction(event -> {
             try{
                 ProdukRepostiory produkRepostiory = new ProdukRepostiory();
-                produkRepostiory.updateProduk(kode_produk.getText(), nama_produk.getText(), kategori.getText(), Integer.parseInt(stock.getText()), Integer.parseInt(harga_modal.getText()), Integer.parseInt(harga_jual.getText()));
+                produkRepostiory.updateProduk(kode_produk.getText(), nama_produk.getText(), kategori.getText(), Integer.parseInt(stock.getText()), Float.parseFloat(harga_modal.getText()), Float.parseFloat(harga_jual.getText()));
                 Back(event);
             }catch (Exception e){
                 System.out.println(e);
